Split download into separate scripts.
All of the download actions are now in /builds/download-<pkg>.sh.
You can download outside of the docker container into a
directory and mount that as /work in Docker.

To see all the downloads ```cat /builds/downloads-*.sh```.
To download all of the packages outside of the docker container.

mkdir work && cd && work
cat /builds/download-*.sh | sh